moved test related methods to test_helper.py
